+++ src/parser/Tokenizer.ts
@@ -377,6 +377,7 @@ DelimiterCloseByOpen[TYPE_OPEN_SYMBOL] = TYPE_CLOSE_SYMBOL;
 DelimiterCloseByOpen[TABLE_OPEN_SYMBOL] = TABLE_CLOSE_SYMBOL;
 DelimiterCloseByOpen[CODE_SYMBOL] = CODE_SYMBOL;
 DelimiterCloseByOpen[DOCS_SYMBOL] = DOCS_SYMBOL;
+DelimiterCloseByOpen[ELISION_SYMBOL] = ELISION_SYMBOL;
 
 export const PairedCloseDelimiters = new Set<string>();
 PairedCloseDelimiters.add(EVAL_CLOSE_SYMBOL);